fixed screen resolution, aspect ratio, some ui, changed player bounds and explosion radius, and implemented PlayerPrefs into GameData
resized screen resolution, aspect ratio, and some UI so it can be seen properly on the itch.io page.

slightly changed player bounds and explosion scale in order to account for the slight screen resolution adjustment.

fixed volume bug where if you started the game on itch.io, the volume would be set to 0. now, it starts at max volume and saves properly.

GameData script now uses PlayerPrefs in order to save data across scenes and sessions.

+++ Assets/Scripts/Game Manager.cs
@@ -44,6 +44,8 @@ public class GameManager : MonoBehaviour
 
     private void Update()
     {
+        // while the player is in the game,
+        // it will keep track of how much the player has played
         if (IsGameActive)
         {
             timePlayed += Time.deltaTime;
@@ -52,6 +54,7 @@ public class GameManager : MonoBehaviour
 
     public void UpdateScore(int incAmount)
     {
+        // updates score in-game
         score += incAmount;
         scoreText.text = "Score: " + score.ToString();
 
@@ -109,7 +112,7 @@ public class GameManager : MonoBehaviour
         DateTime now = DateTime.Now;
 
         GameData.Instance.totalGamesPlayed += 1;
-        GameData.Instance.totalSecondsPlayed += timePlayed;
+        GameData.Instance.totalSecondsPlayed += (int)timePlayed;
 
         GameData.Instance.lastScore = score;
         GameData.Instance.lastScoreDateTime = now.ToString();
